requirements.txt
tqdm
tensorflow-gpu
keras-gpu
numpy
pandas
scikit-learn
matplotlib